Always create cluster in root's PersistentPreRun
All of our commands rely on an ECS cluster called "fargate" -- if it
doesn't exist the user will get confusing error messages. As our goal
is to abstract away the underlying cluster and CreateCluster is an
idempotent operation, we'll hit the endpoint in the PersistentPreRun
to ensure it is always available in the specified region.

- Validate region is "valid" meaning AWS Fargate is available (so, just
  us-east-1)
- Fix bug where AWS_REGION environment variable would be ignored
- Tweak error language when cluster can't be created